backend/requirements.txt
# Основные зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2

# CPU-ONLY PyTorch (без CUDA/GPU поддержки)
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.1.0+cpu
torchvision==0.16.0+cpu